EzyMetrics/ezymetrics/metrics/serializers.py
from rest_framework import serializers
from .models import Lead, Campaign

class LeadSerializer(serializers.ModelSerializer):
    class Meta:
        model = Lead
        fields = '__all__' #including all the fields
class CampaignSerializer(serializers.ModelSerializer):
    class Meta:
        model = Campaign
        fields = '__all__' 
